[#175] Move systemd-related code to a new module
Problem: model.py module is quite huge, also it contains a mix up of
classes required for both native packages and systemd services.

Solution: Move systemd-related classes and functions to a dedicated
module.
